producer/requirements.txt
marshmallow==3.11.1
APScheduler==3.7.0
black==20.8b1
green==3.2.5
requests==2.25.1
kafka-python==2.0.2
pymitter==0.3.0